requirements.txt
streamlit>=1.36.0
tensorflow>=2.12,<3.0
numpy>=1.24
Pillow>=10.0
opencv-python-headless>=4.8
google-generativeai>=0.7.0
